corrected a situation where a canceled order would still be added to the estimate
