+++ src/modules/create-market/components/create-market-form/create-market-form.jsx
@@ -3,6 +3,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
+import { augur } from 'services/augurjs'
 
 import CreateMarketDefine from 'modules/create-market/components/create-market-form-define/create-market-form-define'
 import CreateMarketOutcome from 'modules/create-market/components/create-market-form-outcome/create-market-form-outcome'
@@ -12,6 +13,11 @@ import CreateMarketLiquidityOrders from 'modules/create-market/components/create
 import CreateMarketReview from 'modules/create-market/components/create-market-form-review/create-market-form-review'
 import Styles from 'modules/create-market/components/create-market-form/create-market-form.styles'
 import { ExclamationCircle as InputErrorIcon } from 'modules/common/components/icons'
+import { createBigNumber } from 'utils/create-big-number'
+import { CATEGORICAL, SCALAR } from 'modules/markets/constants/market-types'
+import { BID } from 'modules/transactions/constants/types'
+
+const NEW_ORDER_GAS_ESTIMATE = createBigNumber(700000)
 
 export default class CreateMarketForm extends Component {
 
@@ -50,6 +56,8 @@ export default class CreateMarketForm extends Component {
     this.keyPressed = this.keyPressed.bind(this)
     this.updateState = this.updateState.bind(this)
     this.updateStateValue = this.updateStateValue.bind(this)
+    this.updateInitialLiquidityCosts = this.updateInitialLiquidityCosts.bind(this)
+    this.handleCancelOrder = this.handleCancelOrder.bind(this)
   }
 
   componentWillReceiveProps(nextProps) {
@@ -176,6 +184,68 @@ export default class CreateMarketForm extends Component {
     return validationsArray.every(key => validations[key] === true)
   }
 
+  updateInitialLiquidityCosts(order, shouldReduce) {
+    const {
+      availableEth,
+      newMarket,
+      updateNewMarket,
+    } = this.props
+    const minPrice = newMarket.type === SCALAR ? newMarket.scalarSmallNum : 0
+    const maxPrice = newMarket.type === SCALAR ? newMarket.scalarBigNum : 1
+    const shareBalances = newMarket.outcomes.map(outcome => 0)
+    let outcome
+    let initialLiquidityEth
+    let initialLiquidityGas
+
+    switch (newMarket.type) {
+      case CATEGORICAL:
+        newMarket.outcomes.forEach((outcomeName, index) => {
+          if (this.state.selectedOutcome === outcomeName) outcome = index
+        })
+        break
+      case SCALAR:
+        outcome = this.state.selectedOutcome
+        break
+      default:
+        outcome = 1
+    }
+
+    const orderInfo = {
+      orderType: order.type === BID ? 0 : 1,
+      outcome,
+      shares: order.quantity,
+      price: order.price,
+      tokenBalance: availableEth,
+      minPrice,
+      maxPrice,
+      marketCreatorFeeRate: newMarket.settlementFee,
+      reportingFeeRate: 0,
+      shareBalances,
+      singleOutcomeOrderBook: newMarket.orderBook[outcome] || {},
+    }
+    const action = augur.trading.simulateTrade(orderInfo)
+    // NOTE: Fees are going to always be 0 because we are only opening orders, and there is no costs associated with opening orders other than the escrowed ETH and the gas to put the order up.
+    if (shouldReduce) {
+      initialLiquidityEth = newMarket.initialLiquidityEth.minus(action.tokensDepleted)
+      initialLiquidityGas = newMarket.initialLiquidityGas.minus(NEW_ORDER_GAS_ESTIMATE)
+    } else {
+      initialLiquidityEth = newMarket.initialLiquidityEth.plus(action.tokensDepleted)
+      initialLiquidityGas = newMarket.initialLiquidityGas.plus(NEW_ORDER_GAS_ESTIMATE)
+    }
+
+    updateNewMarket({ initialLiquidityEth, initialLiquidityGas })
+  }
+
+  handleCancelOrder(orderDetails) {
+    const {
+      newMarket,
+      removeOrderFromNewMarket,
+    } = this.props
+    const order = newMarket.orderBook[orderDetails.outcome][orderDetails.index]
+    this.updateInitialLiquidityCosts(order, true)
+    removeOrderFromNewMarket(orderDetails)
+  }
+
   render() {
     const {
       addOrderToNewMarket,
@@ -187,7 +257,6 @@ export default class CreateMarketForm extends Component {
       isMobileSmall,
       meta,
       newMarket,
-      removeOrderFromNewMarket,
       submitNewMarket,
       universe,
       updateNewMarket,
@@ -240,7 +309,7 @@ export default class CreateMarketForm extends Component {
                 updateNewMarket={updateNewMarket}
                 validateNumber={this.validateNumber}
                 addOrderToNewMarket={addOrderToNewMarket}
-                removeOrderFromNewMarket={removeOrderFromNewMarket}
+                updateInitialLiquidityCosts={this.updateInitialLiquidityCosts}
                 availableEth={availableEth}
                 isMobileSmall={isMobileSmall}
                 keyPressed={this.keyPressed}
@@ -297,7 +366,7 @@ export default class CreateMarketForm extends Component {
           { newMarket.currentStep === 3 &&
             <CreateMarketLiquidityOrders
               newMarket={newMarket}
-              removeOrderFromNewMarket={removeOrderFromNewMarket}
+              removeOrderFromNewMarket={this.handleCancelOrder}
               liquidityState={s.liquidityState}
             />
           }
